+++ package.json
@@ -63,6 +63,7 @@
     "@types/blue-tape": "^0.1.32",
     "@types/bluebird": "^3.5.25",
     "@types/body-parser": "^1.17.0",
+    "@types/cheerio": "^0.22.10",
     "@types/cors": "^2.8.4",
     "@types/express": "^4.16.1",
     "@types/express-session": "^1.15.11",
