+++ src/test/request-utils.ts
@@ -1,0 +1,23 @@
+/**
+ * Utility class for test requests
+ */
+export default new class RequestUtils {
+
+  /** 
+   * Creates binary parser for supertest
+  */
+  createBinaryParser() {
+    return (res: any, callback: any) => {
+      res.setEncoding("binary");
+      res.data = "";
+      res.on("data", (chunk: any) => {
+        res.data += chunk;
+      });
+      
+      res.on("end", () => {
+        callback(null, new Buffer(res.data, "binary"));
+      });
+    };
+  }
+
+}
