+++ src/types/pdf2json.d.ts
@@ -1,0 +1,1 @@
+declare module 'pdf2json';
